move resource package location to one place
