+++ resource/configmapresource/desired.go
@@ -1,4 +1,4 @@
-package secret
+package configmapresource
 
 import (
 	"context"
