+++ resource/internal/error.go
